pyproject.toml
[tool.poetry]
name = "stockup"
version = "0.1.0"
description = ""
authors = ["PooriaT <pooria_taghdiri@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.5"
streamlit = "^1.38.0"
google-generativeai = "^0.8.2"
pylint = "^3.3.1"
black = "^24.8.0"
pre-commit = "^3.8.0"
yfinance = "^0.2.44"
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
python-dotenv = "^1.0.1"
plotly = "^5.24.1"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
